pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neural-assemblies"
version = "0.0.1a1"
description = "Neural Assembly Simulation - Assembly Calculus Implementation"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Superintelligent Group", email = "contact@superintelligent.group" },
]
keywords = [
    "neural-assemblies",
    "assembly-calculus",
    "computational-neuroscience",
    "nemo",
    "brain-simulation",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "matplotlib>=3.7.0",
    "pptree>=3.1",
    "tqdm>=4.65.0",
    "pandas>=2.0.0",
    "seaborn>=0.12.0",
]

[project.urls]
Homepage = "https://github.com/Caerii/assemblies"
Documentation = "https://github.com/Caerii/assemblies#readme"
Repository = "https://github.com/Caerii/assemblies"

[project.optional-dependencies]
gpu = [
    "cupy-cuda13x>=13.0.0",
    "torch>=2.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]
all = [
    "neural-assemblies[gpu,dev]",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["neural_assemblies*"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "ipython>=8.0.0",
    "build>=1.0.0",
    "twine>=5.0.0",
]

gpu = [
    "cupy-cuda13x>=13.0.0",
    "torch>=2.0.0",
]

[tool.uv]
[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch" }

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = ["temp/", ".venv/", "*.egg-info", "dist/", "build/"]

[tool.ruff.lint]
# Start with pyflakes (F) errors that catch real bugs.
# E rules are style — enable selectively to avoid noise.
select = [
    "F401",   # unused imports
    "F541",   # f-string without placeholders
    "F811",   # redefined unused name
    "F841",   # unused variable
    "E722",   # bare except
    "E401",   # multiple imports on one line
]
# Ignore import-star issues in __init__.py files and notebooks
per-file-ignores = {"__init__.py" = ["F401"], "*.ipynb" = ["F401", "E402"]}

[tool.pytest.ini_options]
testpaths = ["tests", "neural_assemblies/tests", "research/experiments"]
python_files = ["test_*.py", "*_test.py"]
